Ugly workaround for championgg problems on EC2
For some reason calling the scrapy scraper using a celery worker results
in bizarre selenium errors ('Can not connect to GhostDriver'). Some
searching around points to building phantomjs from source as a possible
fix, but the EC2 instance does not have enough memory to do so. Calling
'scrapy crawl championgg' does not seem to give this error, so we'll do
that for now.