Autocomplete changes and focusout/keydown cleanup
Trying out a different library. Currently there are issues with pressing
tab and sending a request BEFORE the autocompleted value is inserted in
the input field. The request needs to be sent AFTER the value is
inserted.

Cleaned up code dealing with focusout and keydown events on input fields
for summoner names and champion names in index and the PB helper.